{"schema":"decodex/commit/1","summary":"{\"schema\":\"decodex/commit/1\",\"summary\":\"Split elf testkit helper modules\",\"authority\":\"manual\"}","authority":"manual"}
